Task 3B bringup
This script needs a bunch of other commands and scripts running
in addition to Gazebo:
* For the navigation part:
  - Ebot navigation2 stack      (refer Task 1C)
  - Docking service             ( ""   Task 2B)

* For the arm control part:
  - Moveit! controller          ( ""   Task 1B)
  - Servo node trigger request  ( ""   Task 1B)
  - Task 1A needs to be run until the marker TFs are read

To reduce computing load, the script will pause after placing the
rack, and wait for user input before proceeding to picking the boxes.
The user can use this to make sure only the required processes are
running in the background.