Updates after Task 4
Easily the biggest problem encountered during Task 4 was the
unreliability of the ultrasonic sensors. They cannot be used
for docking since they are neither accurate nor precise enough
for the purpose.
Additionally, it was learnt that the robot needs to dock roughly to
the centre of the rack; else, it detaches during navigation (e.g.
in tight curves).

For this reason, the major effort as of now is to get a working docking
solution using IMU; unfortunately, this depends on the rack's exact
orientation being known.

* Ebot docking script:
  - A unified script for both hardware and simulator environments.
    If the __name__ of the module is `ebot_docking_service_hw`, the
    script will behave like it should with the hardware arena; else,
    it will follow the old simulator behaviour.
  - Use IMU for orientation instead of Odometry

* Task 1B: Servo motion at two different speeds
  The arm now slows down as it approaches the destination

* Task 2B: Use AMCL for robot's position in `adjust_pose`
  Hopefully this will help overcome odometry drift

* Task 4C: Preemptively move arm to the upcoming box's shoulder angle

... + other minor changes