laser_utils: improve electromagnet alignment logic
This change introduces a new approach to aligning the ebot's
electromagnets with the rack, after the robot is brought to the
correct location. There are no changes to the positioning logic,
only to the final rotating logic.

In order to rotate to the correct orientation, the target
orientation is calculated after positioning based on
* IMU reading at the time
* Rack's orientation w.r.t the robot at the time

This way, the process depends solely on the IMU sensor's ability
to keep track of orientation *change* during the rotation -- as
opposed to the orientation *value* itself -- and works irrespective
of how much drift the sensor has accumulated.
Additionally, it no longer needs awareness of the rack's expected
orientation. As long as the rack is found by the laser scanner,
the ebot will dock to it.

The only catch to be aware of is that the IMU sensor's reading will
drift if the ebot rotates too slowly.